Move auth endpoints to accounts app; add accounts tests
